add service part 1

+++ src/main/java/Pharmacy/DTO/Response/CategoryResponse.java
@@ -1,7 +1,7 @@
 package Pharmacy.DTO.Response;
 
 public record CategoryResponse(
-        Integer categoryId,
+        Long categoryId,
         String categoryName,
         String categorySlug
 ) {
